# Backend/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: "API Gestion des Transactions"
  description: "API pour gérer les transactions bancaires : création, lecture, mise à jour et suppression."

host: localhost:3001
basePath: /api
schemes:
  - http

paths:
  # Récupérer toutes les transactions
  /transactions:
    get:
      summary: "Récupérer toutes les transactions"
      operationId: getAllTransactions
      responses:
        200:
          description: "Liste de toutes les transactions"
          schema:
            type: array
            items:
              $ref: '#/definitions/Transaction'

    post:
      summary: "Créer une nouvelle transaction"
      operationId: createTransaction
      parameters:
        - in: body
          name: transaction
          description: "Objet transaction à créer"
          required: true
          schema:
            $ref: '#/definitions/TransactionCreate'
      responses:
        201:
          description: "Transaction créée avec succès"
          schema:
            $ref: '#/definitions/Transaction'
        400:
          description: "Champs manquants ou invalides"

  # Récupérer les transactions du mois en cours
  /transactions/current-month:
    get:
      summary: "Récupérer les transactions du mois en cours"
      operationId: getTransactionsCurrentMonth
      parameters:
        - name: accountId
          in: query
          description: "ID du compte (optionnel)"
          required: false
          type: string
      responses:
        200:
          description: "Liste des transactions du mois"
          schema:
            type: array
            items:
              $ref: '#/definitions/Transaction'
        404:
          description: "Compte non trouvé"

  # Détails, mise à jour et suppression d’une transaction spécifique
  /transactions/{transactionId}:
    get:
      summary: "Récupérer une transaction par son ID"
      operationId: getTransactionById
      parameters:
        - name: transactionId
          in: path
          description: "ID de la transaction"
          required: true
          type: string
      responses:
        200:
          description: "Détails de la transaction"
          schema:
            $ref: '#/definitions/Transaction'
        404:
          description: "Transaction non trouvée"

    patch:
      summary: "Modifier une transaction existante"
      operationId: updateTransaction
      parameters:
        - name: transactionId
          in: path
          description: "ID de la transaction à modifier"
          required: true
          type: string
        - in: body
          name: transactionUpdate
          description: "Champs à modifier"
          required: true
          schema:
            $ref: '#/definitions/TransactionUpdate'
      responses:
        200:
          description: "Transaction mise à jour avec succès"
          schema:
            $ref: '#/definitions/Transaction'
        404:
          description: "Transaction non trouvée"

    delete:
      summary: "Supprimer une transaction"
      operationId: deleteTransaction
      parameters:
        - name: transactionId
          in: path
          description: "ID de la transaction à supprimer"
          required: true
          type: string
      responses:
        200:
          description: "Transaction supprimée avec succès"
        404:
          description: "Transaction non trouvée"

definitions:
  Transaction:
    type: object
    required:
      - id
      - amount
      - category
      - date
    properties:
      id:
        type: string
        description: "Identifiant unique de la transaction"
      amount:
        type: number
        description: "Montant de la transaction"
      category:
        type: string
        description: "Catégorie de la transaction"
      date:
        type: string
        format: date-time
        description: "Date de la transaction"

  TransactionCreate:
    type: object
    required:
      - amount
      - category
      - date
    properties:
      amount:
        type: number
        description: "Montant de la nouvelle transaction"
      category:
        type: string
        description: "Catégorie de la nouvelle transaction"
      date:
        type: string
        format: date-time
        description: "Date de la nouvelle transaction"

  TransactionUpdate:
    type: object
    properties:
      amount:
        type: number
        description: "Montant mis à jour"
      category:
        type: string
        description: "Catégorie mise à jour"
      date:
        type: string
        format: date-time
        description: "Date mise à jour"